+++ src/rpcserver.cpp
@@ -5,8 +5,8 @@
 
 #include "rpcserver.h"
 
-#include "base58.h"
 #include "init.h"
+#include "key_io.h"
 #include "random.h"
 #include "sync.h"
 #include "ui_interface.h"
