+++ src/test/base58_tests.cpp
@@ -9,6 +9,7 @@
 #include "data/base58_keys_valid.json.h"
 
 #include "key.h"
+#include "key_io.h"
 #include "script/script.h"
 #include "test/test_bitcoin.h"
 #include "uint256.h"
