+++ src/test/bip32_tests.cpp
@@ -4,8 +4,8 @@
 
 #include <boost/test/unit_test.hpp>
 
-#include "base58.h"
 #include "key.h"
+#include "key_io.h"
 #include "uint256.h"
 #include "util.h"
 #include "utilstrencodings.h"
