+++ src/test/key_tests.cpp
@@ -4,7 +4,7 @@
 
 #include "key.h"
 
-#include "base58.h"
+#include "key_io.h"
 #include "script/script.h"
 #include "uint256.h"
 #include "util.h"
