+++ src/wallet/rpcdump.cpp
@@ -2,7 +2,7 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "base58.h"
+#include "key_io.h"
 #include "rpcserver.h"
 #include "init.h"
 #include "main.h"
