+++ src/wallet/walletdb.cpp
@@ -5,8 +5,8 @@
 
 #include "wallet/walletdb.h"
 
-#include "base58.h"
 #include "consensus/validation.h"
+#include "key_io.h"
 #include "main.h"
 #include "protocol.h"
 #include "serialize.h"
